+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * add rsync client authentication with ssh key
     * add management of redis cluster with sentinel
 * Fixed ...
+    * fixed parse_redis_uri to parse redis sentinel uri
 * Integrated ...
 * Various bugfixes
 * Removed